Single conversation
- One MessageCell class instead of Income* and Outcome*
- thus cellForRowAt refactoring
