--- ChitChat02/Conversation/MessageCell/IncomeMessageCell.swift
@@ -1,38 +1,0 @@
-//
-//  MessageCell.swift
-//  ChitChat02
-//
-//  Created by Timun on 28.09.2020.
-//  Copyright © 2020 TimunInc. All rights reserved.
-//
-
-import UIKit
-
-
-
-class IncomeMessageCell: BaseMessageCell {
-
-    @IBOutlet weak var message: UILabel!
-    @IBOutlet weak var date: UILabel!
-    @IBOutlet weak var background: UIView!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        
-        background.layer.cornerRadius = 8.0
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-    
-}
-
-extension IncomeMessageCell: ConfigurableView {
-    func configure(with model: MessageCellModel) {
-        message.text = model.text
-        date.text = formatDateTime(with: model.date)
-    }
-}
