+++ ChitChat02/Conversation/MessageCell/IncomeMessageCell.xib
@@ -10,7 +10,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="153" id="KGk-i7-Jjw" customClass="IncomeMessageCell" customModule="ChitChat02" customModuleProvider="target">
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="153" id="KGk-i7-Jjw" customClass="MessageCell" customModule="ChitChat02" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="390" height="153"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
@@ -54,9 +54,9 @@
             </tableViewCellContentView>
             <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
             <connections>
-                <outlet property="background" destination="pel-lH-dYB" id="3K4-fM-RFJ"/>
-                <outlet property="date" destination="gC6-1c-mbM" id="8Gn-ba-CzH"/>
-                <outlet property="message" destination="aIx-Ni-hsr" id="znb-YF-99w"/>
+                <outlet property="background" destination="pel-lH-dYB" id="xHY-o6-I6X"/>
+                <outlet property="date" destination="gC6-1c-mbM" id="pc9-th-Q0L"/>
+                <outlet property="message" destination="aIx-Ni-hsr" id="TRi-TP-v6W"/>
             </connections>
             <point key="canvasLocation" x="163.768115942029" y="98.102678571428569"/>
         </tableViewCell>
