+++ ChitChat02/Conversation/MessageCell/MessageCell.swift
@@ -8,26 +8,38 @@
 
 import UIKit
 
-class OutcomeMessageCell: BaseMessageCell {
-
+class MessageCell: UITableViewCell {
+    
     @IBOutlet weak var message: UILabel!
     @IBOutlet weak var date: UILabel!
     @IBOutlet weak var background: UIView!
     
+    private let formatter = DateFormatter()
+    private let calendar = Calendar.current
+
     override func awakeFromNib() {
         super.awakeFromNib()
         
         background.layer.cornerRadius = 8.0
     }
-
+    
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
-
+        
         // Configure the view for the selected state
+    }
+
+    private func formatDateTime(with date: Date) -> String {
+        if calendar.startOfDay(for: date) != calendar.startOfDay(for: Date()) {
+            formatter.dateFormat = "dd MMM"
+        } else {
+            formatter.dateFormat = "HH:mm"
+        }
+        return formatter.string(from: date)
     }
 }
 
-extension OutcomeMessageCell: ConfigurableView {
+extension MessageCell: ConfigurableView {
     func configure(with model: MessageCellModel) {
         message.text = model.text
         date.text = formatDateTime(with: model.date)
