+++ ChitChat02/Conversation/MessageCell/OutcomeMessageCell.xib
@@ -10,7 +10,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="184" id="KGk-i7-Jjw" customClass="OutcomeMessageCell" customModule="ChitChat02" customModuleProvider="target">
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="184" id="KGk-i7-Jjw" customClass="MessageCell" customModule="ChitChat02" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="383" height="184"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
@@ -21,13 +21,13 @@
                         <rect key="frame" x="92" y="4" width="287" height="176"/>
                         <subviews>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Date" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7RJ-d7-MFO">
-                                <rect key="frame" x="247" y="151.5" width="36" height="20.5"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                <nil key="textColor"/>
+                                <rect key="frame" x="258" y="158.5" width="25" height="13.5"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="11"/>
+                                <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="0.25278253424657532" colorSpace="custom" customColorSpace="sRGB"/>
                                 <nil key="highlightedColor"/>
                             </label>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" text="What is the most popular meal in Japan?" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Kiy-18-ceK">
-                                <rect key="frame" x="4" y="4" width="279" height="145.5"/>
+                                <rect key="frame" x="4" y="4" width="279" height="152.5"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
@@ -53,9 +53,9 @@
             </tableViewCellContentView>
             <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
             <connections>
-                <outlet property="background" destination="gdx-Zs-gah" id="ZGb-35-KoC"/>
-                <outlet property="date" destination="7RJ-d7-MFO" id="VsM-lO-fd6"/>
-                <outlet property="message" destination="Kiy-18-ceK" id="WwJ-3O-zuZ"/>
+                <outlet property="background" destination="gdx-Zs-gah" id="ep5-Hf-drL"/>
+                <outlet property="date" destination="7RJ-d7-MFO" id="MdI-9U-tIC"/>
+                <outlet property="message" destination="Kiy-18-ceK" id="FhN-tB-kg6"/>
             </connections>
             <point key="canvasLocation" x="-180" y="75"/>
         </tableViewCell>
